requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
anyio==4.6.0
apache-airflow==2.10.2
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.16.0
apache-airflow-providers-fab==1.3.0
apache-airflow-providers-ftp==3.11.0
apache-airflow-providers-http==4.13.0
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-mysql==5.7.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.6.1
apturl==0.5.2
argcomplete==3.5.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
bcrypt==3.2.0
blinker==1.8.2
Brlapi==0.8.3
cachelib==0.9.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==6.8.2
command-not-found==0.3
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==43.0.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.21.2
duplicity==0.8.21
email_validator==2.2.0
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.16.1
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.9.0
future==0.18.2
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.66.1
gunicorn==23.0.0
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
idna==3.3
importlib_metadata==8.4.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
monotonic==1.6
more-itertools==10.5.0
multidict==6.1.0
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.1.1
netifaces==0.11.0
numpy==2.1.1
oauthlib==3.2.0
olefile==0.46
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
paramiko==2.9.3
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
SecretStorage==3.3.1
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.2
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
ssh-import-id==5.11
systemd-python==234
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
text-unidecode==1.3
texttable==1.6.4
time-machine==2.15.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
universal_pathlib==0.2.5
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.26.5
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
xdg==5
xkit==0.0.0
yarl==1.11.1
zipp==3.20.2
